Add wash_check checks for nigeria flavor

+++ opensrp-chw/src/main/java/org/smartregister/chw/model/WashCheckModel.java
@@ -4,6 +4,7 @@ import org.smartregister.chw.anc.AncLibrary;
 import org.smartregister.chw.anc.domain.Visit;
 import org.smartregister.chw.anc.domain.VisitDetail;
 import org.smartregister.chw.anc.util.NCUtils;
+import org.smartregister.chw.application.ChwApplication;
 import org.smartregister.chw.core.utils.CoreConstants;
 import org.smartregister.chw.dao.WashCheckDao;
 import org.smartregister.clientandeventmodel.Event;
@@ -36,10 +37,12 @@ public class WashCheckModel {
                     }
                 }
             }
-            // delete any previous wash check event
-            List<String> visitIds = WashCheckDao.getLastWashCheckVisitId(familyId);
-            for (int i = 0; i < visitIds.size() - 1; i++) {
-                AncLibrary.getInstance().visitRepository().deleteVisit(visitIds.get(i));
+            if (ChwApplication.getApplicationFlavor().launchWashCheckOnNativeForm()) {
+                // delete any previous wash check event
+                List<String> visitIds = WashCheckDao.getLastWashCheckVisitId(familyId);
+                for (int i = 0; i < visitIds.size() - 1; i++) {
+                    AncLibrary.getInstance().visitRepository().deleteVisit(visitIds.get(i));
+                }
             }
             /*NCUtils.addEvent(allSharedPreferences, baseEvent);
             NCUtils.startClientProcessing();*/
